refactor(external-call): move ExternalCallKey to community/base
Relocate ExternalCallKey and the ExternalCallPayloadDescription payload-size
diagnostics helper from the participant DAMLe engine wrapper into
community/base (com.digitalasset.canton.data). Pure move, no behavior change.

ExternalCallKey is a protocol-data identity (four canonical strings plus an
ordering) derived from ViewParticipantData.ViewExternalCallResult, which
already lives in base. Placing it in base lets TransactionView.validated fail
fast on a recorded-result disagreement in a follow-up -- the centralization
requested in the #552 review (r3506482637) -- and keeps the "same external
call" identity in one place rather than split across DAMLe and the participant
validators.

ExternalCallReplayData and its ReinterpretationError-typed disagreement stay in
DAMLe for now; they move once collapsed to a single output per key in the
fail-fast change.

Refs #565.
