+++ community/base/src/main/scala/com/digitalasset/canton/data/ExternalCallPayloadDescription.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2026 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.digitalasset.canton.participant.util
+package com.digitalasset.canton.data
 
 import com.digitalasset.daml.lf.data.Bytes
